Added files used in reworking rfid code
The program can successfully read hand-typed RFIDs from the command line without a problem but is having issues with reading the input from the card reader. For context, the card reader mimics keystrokes and basically enters the RFID number and then a new line immediately afterward. This was run on a Raspberry Pi 4 using python 2.7.16. The code must be run from the command line with the sudo prefix