Add files via upload
Discovered we were getting correct responses to send to the client but for the wrong reasons. Now checks for an empty list being returned first, then, if not empty, goes into normal checks to see if member is amongst list of authorized users. 

Logging functionality has been finished and will create a new file when needed(Currently monthly) and append in a comma separated format of:
"Granted/Denied, userID, machineIP, date-time"